third_party/blink/web_tests/fast/dom/shadow/shadowroot-clonenode.html

<!DOCTYPE>
<html>
<script src="../../../resources/js-test.js"></script>

<pre id="console"></pre>

<script>
description('Calling ShadowRoot.cloneNode() should throw a NotSupportedError exception.');

var host = document.createElement('div');
var shadowRoot = host.attachShadow({mode: 'open'});
shouldThrow('shadowRoot.cloneNode()', '"NotSupportedError: Failed to execute \'cloneNode\' on \'Node\': ShadowRoot nodes are not clonable."');
</script>
</html>
